latest update
added a getPokemon function that uses the pokemon api to fetch data. i declared variables that were equal to different attributes of pokemon.  i then set the pokedex variable equal to a map of all the attributes that were fetched. we then iterate through the array of all the pokemon and call the get pokemon function within the for loop.